requirements.txt
google-generativeai
pandas
numpy
PyPDF2